backend/requirements.txt
Django==1.11.7
django-cors-headers==2.1.0
django-rest-framework==0.1.0
django-webpack-loader==0.5.0
djangorestframework==3.7.3
lazy-object-proxy==1.3.1
Unidecode==0.4.21